fixup! fixup! fixup! use performance.nodeTiming instead
