+++ test/fixtures/test-nodetiming-uvmetricsinfo.js
@@ -1,0 +1,65 @@
+const { performance } = require('node:perf_hooks');
+const assert = require('node:assert');
+const fs = require('node:fs');
+const { nodeTiming } = performance;
+
+function safeMetricsInfo(cb) {
+  setImmediate(() => {
+    const info = nodeTiming.uvMetricsInfo;
+    cb(info);
+  });
+}
+
+{
+  const info = nodeTiming.uvMetricsInfo;
+  assert.strictEqual(info.loopCount, 0);
+  assert.strictEqual(info.events, 0);
+  // This is the only part of the test that we test events waiting
+  // Adding checks for this property will make the test flaky
+  // as it can be highly influenced by race conditions.
+  assert.strictEqual(info.eventsWaiting, 0);
+}
+
+{
+  // The synchronous call should obviously not affect the uv metrics
+  const fd = fs.openSync(__filename, 'r');
+  fs.readFileSync(fd);
+  const info = nodeTiming.uvMetricsInfo;
+  assert.strictEqual(info.loopCount, 0);
+  assert.strictEqual(info.events, 0);
+  assert.strictEqual(info.eventsWaiting, 0);
+}
+
+{
+  function openFile(info) {
+    assert.strictEqual(info.loopCount, 1);
+    // 1. ? event
+    assert.strictEqual(info.events, 1);
+
+    fs.open(__filename, 'r', (err) => {
+      assert.ifError(err);
+      safeMetricsInfo(afterOpenFile);
+    });
+  }
+
+  function afterOpenFile(info) {
+    assert.strictEqual(info.loopCount, 2);
+    // 1. ? event
+    // 2. uv_fs_open
+    assert.strictEqual(info.events, 2);
+
+    fs.readFile(__filename, (err) => {
+      assert.ifError(err);
+      safeMetricsInfo(afterReadFile);
+    });
+  }
+
+  function afterReadFile(info) {
+    assert.strictEqual(info.loopCount, 6);
+    // 1. ? event
+    assert.strictEqual(info.events, 6);
+    // 1. ?
+  }
+
+  safeMetricsInfo(openFile);
+}
